client/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blyss_vectors"
version = "0.0.1"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "httpx",
    "cryptography",
    "pycose",
    "cbor2",
    "numpy",
    "pyopenssl",
    "pycryptodome"
]
description = "A vector database client that proxies requests through a data obfuscation enclave."
keywords = ["privacy", "vector", "database", "blyss"]

[project.urls]
homepage = "https://blyss.dev"
documentation = "https://docs.blyss.dev"
repository = "https://github.com/blyssprivacy/sap-proxy"
